requirement.txt
asgiref==3.8.1
astroid==3.3.4
black==24.8.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dill==0.3.9
Django==5.1
django-taggit==6.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.8
isort==5.13.2
Markdown==3.7
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
PyJWT==2.9.0
pylint==3.3.1
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
requests==2.32.3
sqlparse==0.5.1
tomlkit==0.13.2
tzdata==2024.1
urllib3==2.2.2
